refactor: use qA instead BaseFragment
