+++ app/src/main/java/pro/sketchware/fragments/settings/events/EventsManagerCreatorFragment.java
@@ -11,20 +11,22 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 
 import com.google.gson.Gson;
+
 import pro.sketchware.R;
 import pro.sketchware.databinding.FragmentEventsManagerCreatorBinding;
+import pro.sketchware.utility.SketchwareUtil;
+import pro.sketchware.utility.FileUtil;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 
-import pro.sketchware.utility.SketchwareUtil;
-import pro.sketchware.utility.FileUtil;
-import pro.sketchware.fragments.base.BaseFragment;
 import mod.hey.studios.util.Helper;
 import mod.hilal.saif.activities.tools.IconSelectorDialog;
 import mod.jbk.util.OldResourceIdMapper;
 
-public class EventsManagerCreatorFragment extends BaseFragment {
+import a.a.a.qA;
+
+public class EventsManagerCreatorFragment extends qA {
 
     private String _code;
     private String _desc;
